status_en.txt: Statutes in English
Translation of the French statutes so anyone can have an idea of how the
association is ran.
It must be said that this is translated by non professional translator,
no one is sworn and as such this has no legal value in France.

The only documents which is accepted by the French state to create such
an association is the status_fr.txt file.

Bear that in mind when reading this text.