accounting/2016: add GSoC-2014 stippends
Buildroot participated in Google Summer of Code 2014, and was awarded
some money for this. There is unfortunately no detailed documents
associated with this transaction, but here are the details:

  - Thomas Petazzoni and Yann E. MORIN mentored a student during the
    GSoC 2014, and Google awarded $500 (€423.14) as stipends for this
    mentoring;

  - Peter Korsgaard and Yann E. MORIN travelled to the GSoC mentors
    reunion in San Jose, for which google re-imbursed $1100 (€930.91)
    for the travel expenses for each participant.

The total amounts to $2700 (€2284.96) and was held by Yann E. MORIN on
his own bank account until there was a bank account for Buildroot. Yann
re-imbursed Peter $1100 (€930.91); Yann kept $100 (€930.91) for himself.

The attached document represents the whole payment by Google to Yann.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Cc: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Cc: Peter Korsgaard <peter@korsgaard.com>