statutes: Change from Foundation to Association
The name could not be 'Buildroot Foundation' as french Law requires a
whole differents sets of rules to declare a foundation, which we are not
going to do.
So, we must change the name to 'Buildroot Association'.

To gain time, the date of the consituent meeting has been changed with
the approval of everyone involved (through mail).